pyproject.toml
[project]
name = "tel-anti"
version = "0.1.0"
description = "Stateful Telegram MCP Server for Antigravity"
authors = [
    { name = "kankinku", email = "your-email@example.com" }
]
dependencies = [
    "mcp[fastmcp]",
    "httpx",
    "python-dotenv",
]
readme = "README.md"
requires-python = ">=3.10"

[project.scripts]
tel-anti = "tel_anti.telegram_mcp_server:mcp.run"

[tool.hatch.build.targets.wheel]
packages = ["src/tel_anti"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
